Change test composition to encapsulate a real composition
On the path to eliminating subtyping for the main composition.
